fixed warning

+++ libs/ftio/srcs/gnl/sfdline.c
@@ -6,7 +6,7 @@
 /*   By: abaurens <abaurens@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/02 15:19:07 by abaurens          #+#    #+#             */
-/*   Updated: 2019/11/14 07:16:35 by abaurens         ###   ########.fr       */
+/*   Updated: 2019/11/14 07:24:21 by abaurens         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,7 +29,7 @@ char		sfdline_(const int fd, char **line, char **save)
 		if (n == 0)
 			break ;
 		buff[n] = 0;
-		if (ft_strlen(buff) < n || !(tmp = gnl_cat_(*save, buff, -1)))
+		if (ft_strlen(buff) < (size_t)n || !(tmp = gnl_cat_(*save, buff, -1)))
 			return (-1);
 		free(*save);
 		*save = tmp;
